master-agent/requirements.txt
schedule==1.2.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0